requirements.txt
django-image-cropping
easy_thumbnails